+++ src/main/java/br/com/zalf/prolog/webservice/commons/util/StringUtils.java
@@ -3,8 +3,6 @@ package br.com.zalf.prolog.webservice.commons.util;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.text.Normalizer;
-
 /**
  * Created on 06/06/18.
  *
@@ -120,28 +118,14 @@ public final class StringUtils {
     }
 
     /**
-     * Return whether the specified string contains only alpabets/special chars too.
-     *
-     * @param string the string to test.
-     * @return true if the string contains only alpabets/special chars, false if it has numbers.
-     */
-    @NotNull
-    public static String stripCharactersSpecials(String string) {
-        String stringFormatada = string.replaceAll("\\-|\\s+", "");
-        return stringFormatada;
-    }
-
-    /**
      * Remove accent.
      *
      * @param string the string to test.
      * @return String without accents.
      */
     @NotNull
-    public static String stripAccents(@NotNull String string) {
-        string = Normalizer.normalize(string, Normalizer.Form.NFD);
-        String stringFormatada = string.replaceAll("[^\\p{ASCII}]", "");
-        return stringFormatada;
+    public static String stripAccents(@NotNull final String string) {
+        return org.apache.commons.lang3.StringUtils.stripAccents(string);
     }
 
     /**