+++ src/main/java/br/com/zalf/prolog/webservice/gente/prontuarioCondutor/model/ocorrencia/Sav.java
@@ -3,8 +3,7 @@ package br.com.zalf.prolog.webservice.gente.prontuarioCondutor.model.ocorrencia;
 /**
  * Created by Zart on 03/07/2017.
  */
-public class Sav {
-
+public final class Sav extends Ocorrencia {
     private int impericia;
     private int imprudencia;